fix(cache): wrong range check for memory
If the addr + 3 equals memorySize in the code, that means the last byte is
right at storageEnd. Therefore, we should change 3 to 4.

In the previous version, the following code will not make any error in ravel:
	.section text
	.globl main
main:
	lw a0, -3(sp)
	ret

But this code has a memory access error since lw from that address should not be valid.